defult bank setiings
