+++ payment/urls.py
@@ -1,6 +1,4 @@
-from django.urls import path, include
-from rest_framework.routers import DefaultRouter
-
+from django.urls import path
 from payment.views import RequestPaymentApi, RequestPaymentVerifyApi
 
 urlpatterns = [
