+++ payment/views.py
@@ -1,19 +1,7 @@
-from django import forms
-from django.shortcuts import render
-
-# Create your views here.
-import zibal.zibal as zibal
-
-import logging
-
-from django.http import Http404
-from django.shortcuts import render
-from django.urls import reverse
 from rest_framework.response import Response
 from rest_framework import serializers
-
-from payment.banks.banks import Zibal
 from rest_framework.views import APIView
+from payment.banks.zibal import Zibal
 
 
 class RequestPaymentApi(APIView):
@@ -33,6 +21,11 @@ class RequestPaymentApi(APIView):
 
 class RequestPaymentVerifyApi(APIView):
     def get(self, request):
+        bank = Zibal()
+        data = {"verify_result": bank.verify(self.request.query_params)}
+        return Response(data)
+
+    def post(self, request):
         bank = Zibal()
         data = {"verify_result": bank.verify(self.request.query_params)}
         return Response(data)
